B file modificaton
The major change here is the fact that now the program looks at every WOCP spectrum in the B File, instead of just normalizing to the first spectrum. This made a very minor difference, but at least we know that laser fluctuation isn't a huge source of error. While I was making this fix, I also cleaned up the code in a bunch of minor ways, mostly removing unnecessary transposes and unused variables.