Slightly better runtime
The new smooth function is not as fast as the old one was, but it's better than it was before. Not sure why it's being so slow, or how such a seemingly messy piece of code can work so much faster. Also, got rid of the CropData function in favor of just having it in the code.